+++ src/App.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import './App.css';
-import UserForm from './components/UserForm';
-import DisplayUser from './components/DisplayUser';
+import UserForm from './components/registration/UserForm';
+import DisplayUser from './components/registration/DisplayUser';
 
 function App() {
 	const [userData, setUserData] = useState(null);
